@@ -1,23 +1,22 @@
 ﻿using Microsoft.Maui.Controls.PlatformConfiguration.AndroidSpecific;
 
-namespace LIN.Vault
+namespace LIN.Vault;
+
+public partial class App : Microsoft.Maui.Controls.Application
 {
-    public partial class App : Microsoft.Maui.Controls.Application
+    public App()
     {
-        public App()
+        InitializeComponent();
+        
+
+        var page = new MainPage()
         {
-            InitializeComponent();
-            
+        };
+        NavigationPage.SetHasNavigationBar(page, false);
+        MainPage = new NavigationPage(page);
 
-            var page = new MainPage()
-            {
-            };
-            NavigationPage.SetHasNavigationBar(page, false);
-            MainPage = new NavigationPage(page);
+        // Ajustar pantalla al teclado.
+        Current?.On<Microsoft.Maui.Controls.PlatformConfiguration.Android>().UseWindowSoftInputModeAdjust(WindowSoftInputModeAdjust.Resize);
 
-            // Ajustar pantalla al teclado.
-            Current?.On<Microsoft.Maui.Controls.PlatformConfiguration.Android>().UseWindowSoftInputModeAdjust(WindowSoftInputModeAdjust.Resize);
-
-        }
     }
 }
